requirements.txt
python-dateutil==2.9.0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2
flask==3.0.3
Werkzeug==3.0.6
pytest==8.3.3
setuptools==75.2.0
flask_cors==5.0.0
fiscalyear==0.4.0
cryptography==41.0.7 # Downgraded to be compatible with snowflake-connector-python 3.7.0
boto3==1.35.49
psycopg2-binary==2.9.9 # For PostgreSQL and Redshift
snowflake-connector-python==3.7.0
google-cloud-storage==2.18.2
azure-storage-blob==12.23.1
azure-identity==1.19.0
waitress==3.0.1